+++ src/pages/List/index.tsx
@@ -1,11 +1,11 @@
 import { Heading, Spinner, Stack } from '@chakra-ui/react'
 import { useIsFetching } from 'react-query'
 
-import { Card } from '../../components/Card'
+import { Card } from '../../components'
 import { USE_WEDDING_GIFTS_KEY } from '../../modules/common/hooks'
 import { Add } from '../../modules/gifts/Add'
 import { ListWeddingGifts } from '../../modules/gifts/List'
-import { Layout } from '../common/layouts/CenterPage'
+import { Layout } from '../common/layouts'
 
 const List = () => {
   const isFetching = useIsFetching([USE_WEDDING_GIFTS_KEY])
