+++ src/pages/common/layouts/index.ts
@@ -1,0 +1,1 @@
+export * from './CenterPage'
